fixed regex defect causing gitlab script error with customer-success config. Also fixed the 'gap
